DEV: Introduce reactive JS API for viewport size
Usage:

```js
@service capabilities;

...

this.capabilities.viewport.sm; // True when viewport is sm or larger
this.capabilities.viewport.md; // True when viewport is md or larger
// etc.
```
These booleans will update live when the browser is resized. Therefore, they should only be used in an autotracking context, so that Ember will re-render things appropriately when they change.